+++ app/game-start/page.tsx
@@ -1,7 +1,7 @@
 import { GameRules } from "@/components/GameRules";
 import { Grid } from "@/components/Grid";
 import * as S from "./styles.css";
-import { GoBackHomeButton } from "@/components/Core/GoBackHomeButton";
+import { GoBackHomeButton } from "@/components/GoBackHomeButton";
 
 export default function GameStartPage() {
   return (
